Update Connect4.js
I pulled out in the main program the functions that were inside other functions. For this, I had to pass more parameters and made some variables global. This was the recommendation received from the Wellcode team. 
Also, I changed the algorithm of counting the consecutive cells of the same colour to determine the winner. With the exception of the down direction, I now count the cells in both sides of the line. In this way, I make sure that in all scenarios, if the number of consecutive cells is equal or bigger than 4, the game is finished.